refactor(groups): move types and update group ID parameter description
