+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./common";
 export * from "./users";
+export * from "./groups";
